requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
pydantic[email]>=2.0.0
httpx>=0.24.1
requests-toolbelt>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0
bcrypt>=4.0.1
asyncpg>=0.28.0  # For PostgreSQL support
aiomysql>=0.2.0  # For MySQL support